backend/requirements.txt
fastapi
uvicorn
openai
sqlalchemy
psycopg2-binary
pandas
python-dotenv
